Authentification

+++ src/facades/species.js
@@ -4,8 +4,8 @@ import formatter from '../utils/formatter';
 
 import config from '../config/config';
 
-const getRecordById = async id => {
-    const speciesRecord = await speciesService.getSpeciesRecordByIdWithFilter(id);
+const getRecordById = async ({ id, accessToken }) => {
+    const speciesRecord = await speciesService.getSpeciesRecordByIdWithFilter({ id, accessToken });
 
     const accepted = formatter.losToTypeaheadSelected(speciesRecord.accepted);
     const basionym = formatter.losToTypeaheadSelected(speciesRecord.basionym);
@@ -36,12 +36,12 @@ const getRecordById = async id => {
     };
 }
 
-const getSpeciesById = async ({ id }) => {
-    return await speciesService.getSpeciesById({ id });
+const getSpeciesById = async ({ id, accessToken }) => {
+    return await speciesService.getSpeciesById({ id, accessToken });
 }
 
-const getAllSpecies = async format => {
-    const listOfSpeciess = await speciesService.getAllSpecies(format);
+const getAllSpecies = async ({ format, accessToken }) => {
+    const listOfSpeciess = await speciesService.getAllSpecies({ format, accessToken });
 
     if (!format) {
         return listOfSpeciess;
@@ -50,8 +50,8 @@ const getAllSpecies = async format => {
     return listOfSpeciess.map(format);
 }
 
-const getAllSpeciesBySearchTerm = async (term, format) => {
-    const listOfSpeciess = await speciesService.getAllSpeciesBySearchTerm(term);
+const getAllSpeciesBySearchTerm = async ({ term, format, accessToken }) => {
+    const listOfSpeciess = await speciesService.getAllSpeciesBySearchTerm({ term, accessToken });
 
     if (!format) {
         return listOfSpeciess;
@@ -59,24 +59,24 @@ const getAllSpeciesBySearchTerm = async (term, format) => {
     return listOfSpeciess.map(format);
 }
 
-const getSynonyms = async (id) => {
+const getSynonyms = async ({ id, accessToken }) => {
 
-    const nomenclatoricSynonyms = await speciesService.getSynonymsNomenclatoricOf({ id });
+    const nomenclatoricSynonyms = await speciesService.getSynonymsNomenclatoricOf({ id, accessToken });
     nomenclatoricSynonyms.sort(helper.listOfSpeciesSorterLex);
 
-    const taxonomicSynonyms = await speciesService.getSynonymsTaxonomicOf({ id });
+    const taxonomicSynonyms = await speciesService.getSynonymsTaxonomicOf({ id, accessToken });
     taxonomicSynonyms.sort(helper.listOfSpeciesSorterLex);
 
-    const invalidDesignations = await speciesService.getInvalidDesignationsOf({ id });
+    const invalidDesignations = await speciesService.getInvalidDesignationsOf({ id, accessToken });
     invalidDesignations.sort(helper.listOfSpeciesSorterLex);
 
     return { nomenclatoricSynonyms, taxonomicSynonyms, invalidDesignations };
 }
 
-const getBasionymsFor = async (id) => {
-    const basionymFor = await speciesService.getBasionymFor({ id });
-    const replacedFor = await speciesService.getReplacedFor({ id });
-    const nomenNovumFor = await speciesService.getNomenNovumFor({ id });
+const getBasionymsFor = async ({ id, accessToken }) => {
+    const basionymFor = await speciesService.getBasionymFor({ id, accessToken });
+    const replacedFor = await speciesService.getReplacedFor({ id, accessToken });
+    const nomenNovumFor = await speciesService.getNomenNovumFor({ id, accessToken });
     return {
         basionymFor,
         replacedFor,
@@ -89,24 +89,26 @@ const saveSpeciesAndSynonyms = async ({
     nomenclatoricSynonyms,
     taxonomicSynonyms,
     invalidDesignations,
+    accessToken,
     isNomenclatoricSynonymsChanged = true,
     isTaxonomicSynonymsChanged = true,
     isInvalidDesignationsChanged = true }) => {
 
-    await speciesService.putNomenclature({ data: species });
+    await speciesService.putNomenclature({ data: species, accessToken });
     await submitSynonyms({
         id: species.id,
         nomenclatoricSynonyms,
         taxonomicSynonyms,
         invalidDesignations,
+        accessToken,
         isNomenclatoricSynonymsChanged,
         isTaxonomicSynonymsChanged,
         isInvalidDesignationsChanged
     });
 }
 
-const saveSpecies = async ({ data }) => {
-    await speciesService.putNomenclature({ data });
+const saveSpecies = async ({ data, accessToken }) => {
+    await speciesService.putNomenclature({ data, accessToken });
 }
 
 async function submitSynonyms({
@@ -114,11 +116,12 @@ async function submitSynonyms({
     nomenclatoricSynonyms,
     taxonomicSynonyms,
     invalidDesignations,
+    accessToken,
     isNomenclatoricSynonymsChanged,
     isTaxonomicSynonymsChanged,
     isInvalidDesignationsChanged }) {
     // get synonyms to be deleted
-    const originalSynonyms = await speciesService.getAllSynonymsOf({ id });
+    const originalSynonyms = await speciesService.getAllSynonymsOf({ id, accessToken });
 
     const toBeDeleted = [];
 
@@ -128,7 +131,8 @@ async function submitSynonyms({
         await saveSynonyms({
             id,
             list: nomenclatoricSynonyms,
-            syntype: config.mappings.synonym.nomenclatoric.numType
+            syntype: config.mappings.synonym.nomenclatoric.numType,
+            accessToken
         });
     }
     if (isTaxonomicSynonymsChanged) {
@@ -136,7 +140,8 @@ async function submitSynonyms({
         await saveSynonyms({
             id,
             list: taxonomicSynonyms,
-            syntype: config.mappings.synonym.taxonomic.numType
+            syntype: config.mappings.synonym.taxonomic.numType,
+            accessToken
         });
     }
     if (isInvalidDesignationsChanged) {
@@ -144,17 +149,18 @@ async function submitSynonyms({
         await saveSynonyms({
             id,
             list: invalidDesignations,
-            syntype: config.mappings.synonym.invalid.numType
+            syntype: config.mappings.synonym.invalid.numType,
+            accessToken
         });
     }
 
     // delete originals
     for (const syn of toBeDeleted) {
-        await speciesService.deleteSynonym({ id: syn.id });
+        await speciesService.deleteSynonym({ id: syn.id, accessToken });
     }
 }
 
-async function saveSynonyms({ id, list, syntype }) {
+async function saveSynonyms({ id, list, syntype, accessToken }) {
     let i = 1;
     for (const s of list) {
         const data = {
@@ -164,7 +170,7 @@ async function saveSynonyms({ id, list, syntype }) {
             rorder: i
         };
         i++;
-        await speciesService.postSynonym({ data });
+        await speciesService.postSynonym({ data, accessToken });
     }
 }
